Remove Plutarch.Convert
